+++ changelog.md
@@ -1,1 +1,1 @@
-- Improved universal reputation
+- Fixed immersive engineering workstation rendering